backend/requirements.txt
flask==2.3.2
python-dotenv==1.0.0
requests==2.31.0
flask-limiter==3.3.1
pytest==7.3.1
gunicorn==20.1.0
